[Feat] SearchCoordinator 로직 추가 & 검색 뷰의 알럿 표시 구현
- SearchViewController와 SearchViewModel 분리
- SearchViewController & SearchViewModel을 Assembly에 등록
- SearchCoordinator 알럿 표시 메서드 & 프로필뷰 이동 메서드 추가
